Added image segmentation for tiling
This code takes in an arbitrary sized full mask image and splits it into seperate 4k(3840x2160) exposures in a directory called tiles.
This folder of seperate exposure tiles will then be used for the tiling exposures. Each tile is named with its relative position in y,x.
So to tile a row of exposures all of the images named tile_row_0 through tile_row_n will be used. Like wise a full column would be
tile_0_column through tile_n_column.
This code is the end of the exposure generation pipeline that starts in KLayout using
parse_whole_exposure to prepare the gds file and gds_to_png_full_design to take the gds of the whole mask and create an properly
sized png based on the number of exposure templements in the mask.
The new split_image_with_overlap does not add any new requirements to the stepper software and can be run using the existing uv script,
but for reference it does make use of PIL image library and the OS library.
The image size and overlap amount are tuneable via the arguments sent to the split_image function but the default values are tuned to
work with the CMU stepper V2.1 and it outputs 4k exposure images with the aligment markers appropratly shared between tiles. However
for other stepper implementations these may need to be tuned especially if a different resolution projector is being used.
Also the directory where the individual tiles are saved can be changed but this shouldn't be needed since the user isn't expected to
directly interact with these images since they should only be used by the stepper software for each tiling exposure.