Added gds2png for full design
This code takes in a gds file that was generated by KLayout and processed by the parse_whole_exposure ruby script in KLayout.
This code is being included in the repo for completeness and a potential integration into the stepper software to enable gds file
compatability, but the current intention is that it will be run in the linked google collab space as the EDA teams gds2png was used.
This code was heavily inspired by the EDA teams gds2png code, but was modified to support exporting an entire mask rather than the
individual templetes. The images created by the non-black background are not needed by the current stepper implementation but where included
to match the functionality of the EDA gds2png code. This code requires the following libraries: gdstk, matplotlib, phidl, pya.
The code leverages the created layer on layer 99 from the parse_whole_exposures ruby script to idenfiy the mask area and to create the full png.
This results in an png that is close to 4k for each template with a 200 pixel overlap between seperate exposures to account for the shared
alignment marks for use in the tiling process. The code can work on any arrangement of templates that are sharing alignment marks and output a
full mask png that can later be split into seperate exposures in the stepper software for a tiling exposure. It also works on a single template
so it could be used as a replacement for the EDA team's gds2png code.