Added exposure parse code for tiling
This code is included in this repo for completeness but it should only be run in KLayout following the EDA team's mask generation SOP
replacing the use of parse_exposures. The code was heavily based on parse_exposures by the EDA team but was adapted to draw the bounding box
around the design area created by all of the templates rather than each template seperatly. This is done so that a full design gds file
is created that can later be converted into a pgn by gds2png(full design) for use in tiling. This code also works on a single template
for more typical designs but the motivation to write another parse_exposures was for tiling. The code draws a bounding box on layer 99
around the full design so that the design area can be indentified and converted into a png. Layer 99 was used to maintain as much compatablity
with any change of layers being used in the HackerFab since no typical design will have that many layers. The code outputs the new gds file
named full_design.gds for use in the gds2png code that is both in this repository and in a google collab space.