Integrated newly segmented images into old tiling flow *UNTESTED*
This code takes the segmented tile images and integrates it into the existing tiling flow.
This is a critical step in moving from the repeated grid exposures of the old implementation to full mask segmentation.
The code also implements the snake pattern of moving through the exposures instead of following the grid left-to-right.
This code is currently being writing in a new TilingFrame towards the bottom of the file but once testing and work is complete it will replace the old TilingFrame.
The old TilingFrame is currently commented out to serve as reference material.
The code has not been tested with the physical stepper but software only tests seem promising with the chip exposure history looking as expected.
A new button has been added that runs the segment images code with being tiling starting the exposure flow. It is required that the Segement Images button be pressed first,
but this requirment is not presently enforced by the software. Also operations that require the use of the camera are currently commented out due to the lack of a software camera to test with.
These issues will be remedied when a tested verison is commited to the repo with this commit facilatiting a test on the stepper.