Temporary Testing File for New Alignment
This file exists to be a testing ground to write modified alignment code that detects and aligns based on two markers on a single edge.
This code is heavily based on the orginal implementation of alignment. Also this code will not replace the orginal implementation of alignment,
but serves a special use exclusivly for tiling specifically, maintaning alignment across exposures.
Once more progress has been made this code will be integrated into the TilingFrame in gui.py but such integration must happen after
the basic tiling flow has been tested and the slight issues referenced in the commit addressed along with any that are discovered through testing.