Minor Revision To Tiling Flow
I was able to preform some dry(no photoresist) tests of the tiling code. It seems that there are some issues with the
autofocus since I am not getting anything close to a repeatable focus escpecailly in UV. Other than this significant
issue the tiling flow worked as expected moving to the appropriate locations and exposing the correct tiles.