Updated Default Moves in X and Y
For this verision I have used the pixel pitch of the DLP471TP and our 10x objective lens to calculate
a better first try for moving the stage between exposures. I found that the width of the projected image should be 1037um and its
height should be about 539um. This should get the chip roughly in place and allow for alignment to do the rest of the work.