Tuned Step Size
The step size between frames has been tuned to reduce the load placed on alignment. At this point the stage movement
alone gets pretty close to having the tiles in the correct place and alignment is able to close these gaps.
There is definetly room for improvement but at this point I am not sure if the source of error is the amount of movement
or error in the stage.